# internal/domains/products/dto/create_product.go
package dto

import "github.com/google/uuid"

type CreateProductDto struct {
	ProductID   uuid.UUID `json:"productId" validate:"required"`
	Name        string    `json:"name" validate:"required,gte=0,lte=255"`
	Description string    `json:"description" validate:"required,gte=0,lte=5000"`
	Price       float64   `json:"price" validate:"required,gte=0"`
}

type CreateProductResponseDto struct {
	ProductID uuid.UUID `json:"productId" validate:"required"`
}